+++ xmake/modules/core/tools/emcc.lua
@@ -1,0 +1,48 @@
+--!A cross-platform build utility based on Lua
+--
+-- Licensed under the Apache License, Version 2.0 (the "License");
+-- you may not use this file except in compliance with the License.
+-- You may obtain a copy of the License at
+--
+--     http://www.apache.org/licenses/LICENSE-2.0
+--
+-- Unless required by applicable law or agreed to in writing, software
+-- distributed under the License is distributed on an "AS IS" BASIS,
+-- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+-- See the License for the specific language governing permissions and
+-- limitations under the License.
+-- 
+-- Copyright (C) 2015-2020, TBOOX Open Source Group.
+--
+-- @author      ruki
+-- @file        emcc.lua
+--
+
+-- inherit gcc
+inherit("gcc")
+
+-- make the strip flag
+function nf_strip(self, level)
+    local maps = 
+    {   
+        debug = "-Wl,-S"
+    ,   all   = "-s"
+    }
+    return maps[level]
+end
+
+-- make the rpathdir flag
+function nf_rpathdir(self, dir)
+end
+
+-- make the link arguments list
+function linkargv(self, objectfiles, targetkind, targetfile, flags, opt)
+
+    -- init arguments
+    opt = opt or {}
+    local argv = table.join("-o", targetfile, objectfiles, flags)
+    if is_host("windows") and not opt.rawargs then
+        argv = winos.cmdargv(argv)
+    end
+    return self:program(), argv
+end